+++ clang-tools-extra/clang-tidy/readability/IdentifierNamingCheck.h
@@ -66,7 +66,7 @@ public:
 
 private:
   llvm::Optional<FailureInfo>
-  GetDeclFailureInfo(const StringRef &Type, const NamedDecl *Decl,
+  getDeclFailureInfo(const StringRef &Type, const NamedDecl *Decl,
                      const SourceManager &SM) const override;
   llvm::Optional<FailureInfo>
   GetMacroFailureInfo(const Token &MacroNameTok,